political_culture/api/messages/serializers.py
from rest_framework import serializers

from political_culture.models import ChatHistory


class ChatHistorySerializer(serializers.ModelSerializer):
    class Meta:
        model = ChatHistory
        fields = ["role", "message"]
