pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = ""
authors = ["fl-korbes <felipekorbes@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
gunicorn = "^23.0.0"
django-environ = "^0.12.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
typedict = "^0.0.4"
mypy = "^1.14.1"
django-stubs = "^5.1.2"
ruff = "^0.9.2"
pytest = "^8.3.4"
hypothesis = "^6.124.2"
langchain = "^0.3.15"
langgraph = "^0.2.67"
langchain-community = "^0.3.15"
tavily-python = "^0.5.0"
langchain-openai = "^0.3.2"
psycopg2-binary = "^2.9.10"
dj-database-url = "^2.3.0"


[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0.20241016"
django-stubs = "^5.1.3"
djangorestframework-stubs = "^3.15.2"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_optional = true
plugins = ["mypy_django_plugin.main", "pydantic.mypy"]


[tool.django-stubs]
django_settings_module = "backend.settings"


select = ["W", "E", "F", "C", "Q"]
target-version = "py312" 
fix = true 
organize-imports = true


[tool.pytest.ini_options]
minversion = "7.0"  # Ensure a specific minimum pytest version
addopts = "-ra -q"  # Custom options for pytest runs
testpaths = [
    "tests"  # Directory where your tests are located
]